+++ lib/runners/sensitive_filter.rb
@@ -21,7 +21,7 @@ module Runners
         source = @options.source
         if source.is_a?(Options::GitSource)
           # @type var source: Options::GitSource
-          user_info = source.git_http_userinfo
+          user_info = source.git_url_userinfo
           list << user_info if user_info
         end
         list