+++ test/test_helper.rb
@@ -35,9 +35,7 @@ module TestHelper
   def new_source(**source)
     {
       head: "6ba85479fc406a64b7202f7bc3ea8da3ada93084",
-      git_http_url: "https://github.com",
-      owner: "sider",
-      repo: "runners_test",
+      git_url: "https://github.com/sider/runners_test",
       **source,
     }
   end
@@ -77,10 +75,8 @@ module TestHelper
     source = {
       head: "123abc",
       base: "456def",
-      git_http_url: "https://github.com",
-      owner: "foo",
-      repo: "bar",
-      git_http_userinfo: "user:secret",
+      git_url: "https://github.com/foo/bar",
+      git_url_userinfo: "user:secret",
       pull_number: 105,
     }
     with_runners_options_env(source: source) do